Refactor: Master/Slave 관련 로직 재정리 및 위치 이동 그리고 yml 파일 분리해라 - #46
