--- tictoc-api/src/main/java/tictoc/config/WebConfig.java
@@ -1,19 +1,0 @@
-package tictoc.config;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import tictoc.config.security.resolver.UserIdArgumentResolver;
-import java.util.List;
-
-@RequiredArgsConstructor
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-    private final UserIdArgumentResolver userIdArgumentResolver;
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(userIdArgumentResolver);
-    }
-}
