--- tictoc-api/src/main/java/tictoc/config/security/exception/JwtAuthenticationEntryPoint.java
@@ -1,40 +1,0 @@
-package tictoc.config.security.exception;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-import tictoc.constants.AuthConstants;
-import tictoc.error.ErrorCode;
-import tictoc.error.ErrorResponse;
-
-import java.io.IOException;
-import java.io.PrintWriter;
-
-@RequiredArgsConstructor
-@Component
-public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
-    private final ObjectMapper objectMapper;
-
-    @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException {
-        handleException(response);
-    }
-
-    private void handleException(HttpServletResponse response) throws IOException {
-        setResponse(response, HttpStatus.UNAUTHORIZED, ErrorCode.UNAUTHORIZED);
-    }
-
-    private void setResponse(HttpServletResponse response, HttpStatus httpStatus, ErrorCode failureCode) throws IOException {
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.setCharacterEncoding(AuthConstants.CHARACTER_TYPE);
-        response.setStatus(httpStatus.value());
-        PrintWriter writer = response.getWriter();
-        writer.write(objectMapper.writeValueAsString(ErrorResponse.of(failureCode)));
-    }
-}
