--- tictoc-api/src/main/java/tictoc/config/security/jwt/JwtProperties.java
@@ -1,15 +1,0 @@
-package tictoc.config.security.jwt;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Getter
-@Setter
-@ConfigurationProperties("jwt")
-@Component
-public class JwtProperties {
-    private String secret;
-    private long accessTokenExpireTime;
-}
