--- tictoc-api/src/main/java/tictoc/config/security/jwt/repository/RefreshTokenRepository.java
@@ -1,9 +1,0 @@
-package tictoc.config.security.jwt.repository;
-
-import org.springframework.data.repository.CrudRepository;
-import tictoc.config.security.jwt.RefreshToken;
-
-public interface RefreshTokenRepository extends CrudRepository<RefreshToken, String> {
-    void deleteByUserId(final Long userId);
-    RefreshToken findByToken(String token);
-}
