+++ tictoc-api/src/main/java/tictoc/database/security/filter/TokenAuthentication.java
@@ -1,0 +1,54 @@
+package tictoc.database.security.filter;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+
+import java.util.Collection;
+import java.util.Collections;
+
+@RequiredArgsConstructor
+public class TokenAuthentication implements Authentication {
+    private final String token;
+    private final Long userId;
+    private boolean isAuthenticated = true;
+
+    @Override
+    public String getName() {
+        return String.valueOf(userId);
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return Collections.emptyList();
+    }
+
+    @Override
+    public String getCredentials() {
+        return token;
+    }
+
+    @Override
+    public Object getDetails() {
+        return null;
+    }
+
+    @Override
+    public Long getPrincipal() {
+        return userId;
+    }
+
+    @Override
+    public boolean isAuthenticated() {
+        return isAuthenticated;
+    }
+
+    @Override
+    public void setAuthenticated(boolean isAuthenticated) {
+        this.isAuthenticated = isAuthenticated;
+    }
+
+    public static TokenAuthentication createTokenAuthentication(final String token, final long userId) {
+        return new TokenAuthentication(token, userId);
+    }
+}
