+++ tictoc-api/src/main/java/tictoc/database/security/jwt/model/Token.java
@@ -1,0 +1,17 @@
+package tictoc.database.security.jwt.model;
+
+import lombok.AccessLevel;
+import lombok.Builder;
+
+@Builder(access = AccessLevel.PRIVATE)
+public record Token(
+        String accessToken,
+        String refreshToken
+) {
+    public static Token of(final String accessToken, final String refreshToken) {
+        return Token.builder()
+                .accessToken(accessToken)
+                .refreshToken(refreshToken)
+                .build();
+    }
+}
