+++ tictoc-api/src/main/java/tictoc/user/port/in/UserCommandUseCase.java
@@ -1,6 +1,6 @@
 package tictoc.user.port.in;
 
-import tictoc.config.security.jwt.dto.JwtResDTO;
+import tictoc.database.security.jwt.dto.JwtResDTO;
 import tictoc.user.dto.request.UserUseCaseReqDTO;
 
 public interface UserCommandUseCase {
