+++ tictoc-api/src/main/java/tictoc/user/service/UserCommandService.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import tictoc.error.ErrorCode;
-import tictoc.config.security.jwt.dto.JwtResDTO;
-import tictoc.config.security.jwt.util.JwtProvider;
+import tictoc.database.security.jwt.dto.JwtResDTO;
+import tictoc.database.security.jwt.util.JwtProvider;
 import tictoc.user.dto.request.UserUseCaseReqDTO;
 import tictoc.user.port.in.UserCommandUseCase;
 import tictoc.user.port.out.UserAdaptor;
