+++ tictoc-common/src/main/java/tictoc/config/database/MasterDataSourceConfig.java
@@ -1,0 +1,22 @@
+package tictoc.config.database;
+
+import com.zaxxer.hikari.HikariDataSource;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.jdbc.DataSourceBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+import javax.sql.DataSource;
+
+@Configuration
+public class MasterDataSourceConfig {
+    @Primary
+    @Bean(name = "masterDataSource")
+    @ConfigurationProperties(prefix="spring.datasource.master.hikari")
+    public DataSource masterDataSource() {
+        return DataSourceBuilder.create()
+                .type(HikariDataSource.class)
+                .build();
+    }
+
+}
