+++ tictoc-common/src/main/java/tictoc/config/database/ReplicationRoutingDataSource.java
@@ -1,0 +1,14 @@
+package tictoc.config.database;
+
+import org.springframework.jdbc.datasource.lookup.AbstractRoutingDataSource;
+import org.springframework.transaction.support.TransactionSynchronizationManager;
+import tictoc.config.database.type.DataSourceType;
+
+public class ReplicationRoutingDataSource extends AbstractRoutingDataSource {
+
+    @Override
+    protected Object determineCurrentLookupKey() {
+        DataSourceType dataSourceType = TransactionSynchronizationManager.isCurrentTransactionReadOnly() ? DataSourceType.Slave : DataSourceType.Master;
+        return dataSourceType;
+    }
+}
