+++ tictoc-common/src/main/java/tictoc/config/database/SlaveDataSourceConfig.java
@@ -1,0 +1,20 @@
+package tictoc.config.database;
+
+import com.zaxxer.hikari.HikariDataSource;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.jdbc.DataSourceBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import javax.sql.DataSource;
+
+@Configuration
+public class SlaveDataSourceConfig {
+    @Bean(name = "slaveDataSource")
+    @ConfigurationProperties(prefix="spring.datasource.slave.hikari")
+    public DataSource slaveDataSource() {
+        return DataSourceBuilder.create()
+                .type(HikariDataSource.class)
+                .build();
+    }
+}
