+++ tictoc-common/src/main/java/tictoc/config/database/type/DataSourceType.java
@@ -1,0 +1,5 @@
+package tictoc.config.database.type;
+
+public enum DataSourceType {
+    Master, Slave
+}
