--- tictoc-domain/src/main/java/tictoc/config/MasterDataSourceConfig.java
@@ -1,24 +1,0 @@
-package tictoc.config;
-
-import com.zaxxer.hikari.HikariDataSource;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-import javax.sql.DataSource;
-
-@Configuration
-@Slf4j
-public class MasterDataSourceConfig {
-    @Primary
-    @Bean(name = "masterDataSource")
-    @ConfigurationProperties(prefix="spring.datasource.master.hikari")
-    public DataSource masterDataSource() {
-        return DataSourceBuilder.create()
-                .type(HikariDataSource.class)
-                .build();
-    }
-
-}
