--- tictoc-domain/src/main/java/tictoc/config/ReplicationRoutingDataSource.java
@@ -1,16 +1,0 @@
-package tictoc.config;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.jdbc.datasource.lookup.AbstractRoutingDataSource;
-import org.springframework.transaction.support.TransactionSynchronizationManager;
-import tictoc.config.type.DataSourceType;
-
-@Slf4j
-public class ReplicationRoutingDataSource extends AbstractRoutingDataSource {
-
-    @Override
-    protected Object determineCurrentLookupKey() {
-        DataSourceType dataSourceType = TransactionSynchronizationManager.isCurrentTransactionReadOnly() ? DataSourceType.Slave : DataSourceType.Master;
-        return dataSourceType;
-    }
-}
