--- tictoc-domain/src/main/java/tictoc/config/SlaveDataSourceConfig.java
@@ -1,24 +1,0 @@
-package tictoc.config;
-
-import com.zaxxer.hikari.HikariDataSource;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import javax.sql.DataSource;
-
-@Configuration
-@Slf4j
-public class SlaveDataSourceConfig {
-
-    @Bean(name = "slaveDataSource")
-    @ConfigurationProperties(prefix="spring.datasource.slave.hikari")
-    public DataSource slaveDataSource() {
-        return DataSourceBuilder.create()
-                .type(HikariDataSource.class)
-                .build();
-    }
-
-}
