--- tictoc-domain/src/main/java/tictoc/config/type/DataSourceType.java
@@ -1,5 +1,0 @@
-package tictoc.config.type;
-
-public enum DataSourceType {
-    Master, Slave
-}
