+++ tictoc-domain/src/main/resources/application-domain.yml
@@ -1,29 +1,6 @@
 spring:
   config:
     import: optional:file:.env[.properties]
-  main:
-    allow-bean-definition-overriding: true
-  datasource:
-    master:
-      hikari:
-        driver-class-name: com.mysql.cj.jdbc.Driver
-        jdbc-url: jdbc:mysql://${MASTER_DB_HOST}:${MASTER_DB_PORT}/${MASTER_DB_NAME}?allowPublicKeyRetrieval=true&useSSL=false&serverTimezone=Asia/Seoul&characterEncoding=UTF-8
-        username: ${DB_USER}
-        password: ${DB_PASSWORD}
-
-    slave:
-      hikari:
-        driver-class-name: com.mysql.cj.jdbc.Driver
-        jdbc-url: jdbc:mysql://${SLAVE_DB_HOST}:${SLAVE_DB_PORT}/${SLAVE_DB_NAME}?allowPublicKeyRetrieval=true&useSSL=false&serverTimezone=Asia/Seoul&characterEncoding=UTF-8
-        username: ${DB_USER}
-        password: ${DB_PASSWORD}
-#
-#    hikari:
-#      connection-timeout: 50000
-#      maximum-pool-size: 100
-#      idle-timeout: 300000
-#      max-lifetime: 1800000
-
   jpa:
     show-sql: true
     properties:
